compare_generic_iter():  Fixed the failure of test_wsgiref's testFileWrapper when running with -O.
test_simple_validation_error still fails under -O.  That appears to be because
wsgiref's validate.py uses `assert` statements all over the place to check
arguments for sanity.  That should all be changed (it's not a logical error
in the software if a user passes bogus arguments, so this isn't a reasonable
use for `assert` -- checking external preconditions should generally raise
ValueError or TypeError instead, as appropriate).

--HG--
branch : legacy-trunk