Reorganize the docs for 'file' and 'open()' after some discussion with Fred.
We want to encourage users to write open() when opening a file, but
open() was described with a single paragraph and
'file' had lots of explanation of the mode and bufsize arguments.

I've shrunk the description of 'file' to cross-reference to the 'File
objects' section, and to open() for an explanation of the arguments.

open() now has all the paragraphs about the mode string.  The bufsize
argument was moved up so that it isn't buried at the end; now there's
1 paragraph on mode, 1 on bufsize, and then 3 more on mode.  Various
other edits and rearrangements were made in the process.

It's probably best to read the final text and not to try to make sense
of the diffs.

--HG--
branch : legacy-trunk