Update the encoding package's search function to use absolute imports when calling __import__.  This helps make the expected search locations for encoding modules be more explicit.
One could use an explicit value for __path__ when making the call to __import__
to force the exact location searched for encodings.  This would give the most
strict search path possible if one is worried about malicious code being
imported.  The unfortunate side-effect of that is that if __path__ was modified
on 'encodings' on purpose in a safe way it would not be picked up in future
__import__ calls.

--HG--
branch : legacy-trunk